♻️refactor: 회식 상세 컴포넌트 폴더 정리
