+++ src/components/dining/common/dining-participant-list.tsx
