+++ src/components/dining/common/dining-status-badge.tsx
