+++ src/components/dining/common/index.ts
