+++ src/components/dining/restaurant-vote/restaurant-vote-fallback.tsx
